requirements.docker.txt
# Core FastAPI & async
fastapi==0.115.9
uvicorn[standard]==0.24.0
websockets==12.0
httpx>=0.25.0
requests>=2.31.0
typing-extensions>=4.8.0

# Database & ORM
asyncpg==0.29.0
sqlalchemy==2.0.27
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1   # satisfies embedchain's ≥1.13.1
databases[postgresql]==0.9.0

psycopg2-binary==2.9.9

# Vector DB
qdrant-client>=1.7.0

# Embeddings for semantic search
sentence-transformers>=2.2.0

# Vector embeddings for semantic search
sentence-transformers>=2.2.0

# CrewAI (use latest compatible version without dependency conflicts)
crewai==0.130.0

# OpenAI and LLM
openai==1.68.2

# Google AI (for Gemini support)
google-generativeai>=0.3.0

# Anthropic (for Claude support)
anthropic>=0.7.0

# (No langchain, langchain-openai, langchain-community, langchain-core, langfuse removed)

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP & async utilities
httpx==0.27.0
aiofiles==23.2.1
redis>=4.5.4  # Modern redis with async support

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.7.4
python-multipart==0.0.6
beautifulsoup4==4.12.2

# Monitoring & logging
prometheus_client==0.19.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
pathlib2==2.3.7
click==8.1.7
beautifulsoup4>=4.12.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0

# Tokenizers
tiktoken>=0.5.2,<1.0.0
